document_analysis/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.21
openai>=1.3.0
tiktoken>=0.5.2
python-dotenv>=1.0.0

# NLP and Text Processing
spacy>=3.7.2
nltk>=3.8.1
textblob>=0.17.1

# Vector Store
faiss-cpu>=1.7.4
numpy>=1.21.0
scikit-learn>=1.4.0

# Web Interface
streamlit>=1.32.0

# Testing and Development
pytest>=8.0.0
pytest-cov>=4.1.0
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0

# Download spaCy model
# python -m spacy download en_core_web_sm
